+++ NAMESPACE
@@ -23,6 +23,7 @@ export(change_gene_identifier)
 export(community_detection)
 export(contrastive_pca_processing)
 export(cor_module_check_epsilon)
+export(cor_module_coremo_clustering)
 export(cor_module_graph_check_res)
 export(cor_module_graph_final_modules)
 export(cor_module_processing)
