+++ man/scale_free_fit.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/methods_coexp_cor.R
-\name{.scale_free_fit}
-\alias{.scale_free_fit}
+\name{scale_free_fit}
+\alias{scale_free_fit}
 \title{Calculate the goodness of fit for a power law distribution.}
 \usage{
-.scale_free_fit(k, breaks = 50L, plot = FALSE)
+scale_free_fit(k, breaks = 50L, plot = FALSE)
 }
 \arguments{
 \item{k}{Numeric vector. The vector of node degrees.}
